Reduce count of index.js Fixes #118
