+++ frontend/src/components/HeroRow.js
