+++ frontend/src/components/UserPage.js
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { restoreSession } from "../../store/session";
+import { restoreSession } from "../store/session";
 import NotesList from "./NotesList";
 
 export default function UserPage() {
